+++ Domain/src/main/java/com/sdk/domain/entity/asset/AccessFeeEntity.kt
@@ -1,6 +1,6 @@
 package com.sdk.domain.entity.asset
 
 
-data class AccessFeeEntity(val id: Long, val merchantId: Long, val amount: Long, val currency: String, val description: String,
+data class AccessFeeEntity(val id: Long, val merchantId: Long, val amount: Float, val currency: String, val description: String,
                            val accessTypeEntity: AccessTypeEntity, val itemType: String, val trialPeriodEntity: TrialPeriodEntity?, val setupFeeEntity: SetupFeeEntity?,
                            val expiresAt: Long?)
